+++ modules/mvt/src/mvt-loader.ts
@@ -1,9 +1,20 @@
 import type {Loader, LoaderWithParser} from '@loaders.gl/loader-utils';
+import type {MVTLoaderOptions} from './lib/types';
 import parseMVT from './lib/parse-mvt';
 
 // __VERSION__ is injected by babel-plugin-version-inline
 // @ts-ignore TS2304: Cannot find name '__VERSION__'.
 const VERSION = typeof __VERSION__ !== 'undefined' ? __VERSION__ : 'latest';
+
+const DEFAULT_MVT_LOADER_OPTIONS: MVTLoaderOptions = {
+  mvt: {
+    shape: 'geojson',
+    coordinates: 'local',
+    layerProperty: 'layerName',
+    layers: undefined,
+    tileIndex: null
+  }
+};
 
 /**
  * Worker loader for the Mapbox Vector Tile format
@@ -22,14 +33,7 @@ export const MVTWorkerLoader: Loader = {
   ],
   worker: true,
   category: 'geometry',
-  options: {
-    mvt: {
-      coordinates: 'local',
-      layerProperty: 'layerName',
-      layers: null,
-      tileIndex: null
-    }
-  }
+  options: DEFAULT_MVT_LOADER_OPTIONS
 };
 
 /**
@@ -37,7 +41,7 @@ export const MVTWorkerLoader: Loader = {
  */
 export const MVTLoader: LoaderWithParser = {
   ...MVTWorkerLoader,
-  parse: async (arrayBuffer, options) => parseMVT(arrayBuffer, options),
+  parse: async (arrayBuffer, options?: MVTLoaderOptions) => parseMVT(arrayBuffer, options),
   parseSync: parseMVT,
   binary: true
 };