+++ modules/schema/src/index.ts
@@ -8,13 +8,16 @@ export type {
   Table,
   ArrayRowTable,
   ObjectRowTable,
+  GeoJSONRowTable,
   ColumnarTable,
-  ArrowTable
+  ArrowTable,
+  Tables
 } from './category/table/table-types';
 export type {
   TableBatch,
   RowArrayTableBatch,
   RowObjectTableBatch,
+  GeoJSONRowTableBatch,
   ColumnarTableBatch,
   ArrowTableBatch
 } from './category/table/table-types';
@@ -53,8 +56,13 @@ export type {ImageDataType, ImageType, ImageTypeEnum} from './category/image/ima
 
 // TYPES
 // GIS CATEGORY - GEOJSON
-export type {GeoJSON, Feature, Geometry, Position, GeoJsonProperties} from './category/gis';
 export type {
+  GeoJSON,
+  Feature,
+  FeatureCollection,
+  Geometry,
+  Position,
+  GeoJsonProperties,
   Point,
   MultiPoint,
   LineString,