+++ modules/wkt/src/wkb-writer.ts
@@ -3,7 +3,7 @@ import {VERSION} from './lib/utils/version';
 import encodeWKB from './lib/encode-wkb';
 
 /**
- * WKT exporter
+ * WKB exporter
  */
 export const WKBWriter: Writer = {
   name: 'WKB (Well Known Binary)',
@@ -14,6 +14,6 @@ export const WKBWriter: Writer = {
   // @ts-ignore
   encode: encodeWKB,
   options: {
-    wkt: {}
+    wkb: {}
   }
 };